% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cehmet-package.r, R/cehmet.R
\docType{package}
\name{cehmet}
\alias{cehmet}
\alias{cehmet-package}
\alias{cehmet}
\title{cehmet.}
\usage{
cehmet(cehmet, dest, public = FALSE, ...)
}
\arguments{
\item{cehmet}{url to the cehmet repository}

\item{dest}{destination remote url}

\item{public}{should the new GitHub repository be public or private?}

\item{...}{additional arguments to git2r functions}
}
\description{
cehmet.

Clone a cehmet, change remote origin, push
}
\details{
url can be any git address we can use to clone (e.g. https://, git@github.com:)
}
\examples{
\dontrun{
cehmet("https://github.com/cboettig/cehmet",
         "https://gtihub.com/boettiger-lab/mypkg")
}
}
